Added vmIfAndSwitch_test.go
